added Update Post Feature

+++ frontend/src/pages/users/Dashboard.jsx
@@ -24,14 +24,18 @@ const Dashboard = () => {
 
   // Handle Delete Post
   const handleDelete = async (_id) => {
-    try {
-      const data = await deletePost(_id);
-      setSuccess(data.success);
-    } catch (error) {
-      setError(error.message)
+
+    if (confirm("Confirm delete?")) {
+      try {
+        const data = await deletePost(_id);
+        setSuccess(data.success);
+      } catch (error) {
+        setError(error.message)
+      }
+      const newPosts = user.posts.filter(post => post._id !== _id)
+      setUser({...user, posts: newPosts})
     }
-    const newPosts = user.posts.filter(post => post._id !== _id)
-    setUser({...user, posts: newPosts})
+
   }
 
   useEffect(() => {
@@ -62,6 +66,7 @@ const Dashboard = () => {
           <div className="flex items-center gap-2">
             <Link className="fa-solid fa-pen-to-square nav-link text-green-500 hover:bg-green-200"
             title="Update"
+            state={post}
             to="/update"
             ></Link>
             <button className="fa-solid fa-trash-can nav-link text-red-500 hover:bg-red-200"
